wallet-ext: address pr comments
