+++ apps/wallet/src/ui/app/hooks/useGetObject.ts
@@ -2,31 +2,25 @@
 // SPDX-License-Identifier: Apache-2.0
 
 import { useRpcClient } from '@mysten/core';
-import {
-    type SuiObjectResponse,
-    normalizeSuiAddress,
-    type SuiObjectDataOptions,
-} from '@mysten/sui.js';
+import { type SuiObjectResponse, normalizeSuiAddress } from '@mysten/sui.js';
 import { useQuery, type UseQueryResult } from '@tanstack/react-query';
 
 export function useGetObject(
-    objectId: string,
-    options?: SuiObjectDataOptions
+    objectId: string
 ): UseQueryResult<SuiObjectResponse, unknown> {
-    const adjOptions = {
-        showType: true,
-        showContent: true,
-        showOwner: true,
-        ...options,
-    };
     const rpc = useRpcClient();
     const normalizedObjId = normalizeSuiAddress(objectId);
     const response = useQuery(
-        ['object', normalizedObjId, adjOptions],
+        ['object', normalizedObjId],
         async () => {
             return rpc.getObject({
                 id: normalizedObjId,
-                options: adjOptions,
+                options: {
+                    showType: true,
+                    showContent: true,
+                    showOwner: true,
+                    showDisplay: true,
+                },
             });
         },
         { enabled: !!objectId }
