+++ apps/wallet/src/ui/app/hooks/useOwnedNFT.ts
@@ -15,7 +15,7 @@ export function useOwnedNFT(
     nftObjectId: string | null,
     address: SuiAddress | null
 ) {
-    const data = useGetObject(nftObjectId || '', { showDisplay: true });
+    const data = useGetObject(nftObjectId || '');
     const { data: objectData } = data;
     const objectDetails = useMemo(() => {
         if (!objectData || !is(objectData.data, SuiObjectData) || !address)
